requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.32.0
APScheduler==3.10.4
